// packages/bigcommerce/src/types/products_catalog.v3.d.ts
export interface paths {
  "/catalog/products": {
    /**
     * Get All Products
     * @description Returns a list of **Products**. Optional filter parameters can be passed in.
     */
    get: operations["getProducts"];
    /**
     * Update Products (Batch)
     * @description Updates products in batches. Batches are limited to 10 products.
     *
     * **Required Fields**
     * * `id` - product `id` is required for batch updates to products.
     *
     * **Read-Only Fields**
     * - `id`
     * - `date_created`
     * - `date_modified`
     * - `calculated_price`
     * - `base_variant_id`
     */
    put: operations["updateProducts"];
    /**
     * Create a Product
     * @description Creates a *Product*. Only one product can be created at a time; however, you can create multiple product variants using the `variants` array.
     *
     * **Required Fields:**
     * - `name`
     * - `type`
     * - `weight`
     * - `price`
     *
     * **Read-Only Fields**
     * - `id`
     * - `date_created`
     * - `date_modified`
     * - `calculated_price`
     * - `base_variant_id`
     *
     * **Limits**
     * - 250 characters product name length.
     * - A product can have up to 1000 images. Each image file or image uploaded by URL can be up to 8 MB.
     *
     * **Usage Notes**
     * * You can create multiple product variants using the `variants` array.
     * * This endpoint accepts a `video` array. To create a product video that accepts a `video` object, see [Create a Product Video](/docs/rest-catalog/products/videos#create-a-product-video) for information.
     */
    post: operations["createProduct"];
    /**
     * Delete Products
     * @description To delete *Product* objects, you must include a filter. This prevents inadvertently deleting all *Product* objects in a store.
     *
     * > #### Note
     * > The maximum number of products you can delete at one time is 250.
     *
     * **Example**:
     * To delete products with IDs 1,2 and 3, use `DELETE /v3/catalog/products?id:in=1,2,3`.
     */
    delete: operations["deleteProducts"];
    parameters: {
      header: {
        Accept: components["parameters"]["Accept"];
      };
    };
  };
  "/catalog/products/{product_id}": {
    /**
     * Get a Product
     * @description Returns a single *Product*. Optional parameters can be passed in.
     */
    get: operations["getProduct"];
    /**
     * Update a Product
     * @description Updates a *Product*.
     *
     * **Limits**
     * - A product can have up to 1000 images. Each image file or image uploaded by URL can be up to 8 MB.
     *
     * **Read-Only Fields**
     * - id
     * - date_created
     * - date_modified
     * - calculated_price
     * - base_variant_id
     */
    put: operations["updateProduct"];
    /**
     * Delete a Product
     * @description Deletes a *Product*.
     */
    delete: operations["deleteProduct"];
    parameters: {
      header: {
        Accept: components["parameters"]["Accept"];
      };
      path: {
        product_id: components["parameters"]["ProductIdParam"];
      };
    };
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    MetaFieldCollectionResponse: {
      data?: components["schemas"]["Metafield"][];
      meta?: components["schemas"]["CollectionMeta"];
    };
// ...
  };
// ...
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {

  /**
   * Get All Products
   * @description Returns a list of **Products**. Optional filter parameters can be passed in.
   */
  getProducts: {
// ...
    responses: {
      200: {
        content: {
          "application/json": {
            data?: components["schemas"]["product_Full"][];
            meta?: components["schemas"]["metaCollection_Full"];
          };
        };
      };
    };
  };
// ...
}